Version 0.2.0; reworked the OSC Input node.
The OSC Input node is now the On OSC Message node, and provides a
flow output that is triggered any time a message with the given address
is received, as well as allowing for configuration of any number of
(simple) typed arguments. This makes the node both more versatile and
simpler to use.

This is also a breaking change from the previous release, and any
instances of the old node will disappear as the node's ID has changed.
This plugin is still not published on the Steam workshop, so I consider
this acceptable for now, but will refrain from breaking changes where
possible once it's published.