Refined data output port management.
This causes only the changed ports to be updated, rather than removing
and recreating every port every time an input argument changes.
Functionally, this doesn't change much.

There are still some odd issues remaining if you change the type of a
port while it's connected to other nodes, but there doesn't seem to be
a good solution to that. All connections can be removed by scanning all
of the connections in the graph and removing the corresponding ones, but
the editor still doesn't update to remove the connection and an error
occurs if you try to remove the ghost connection.

I'd like to fix it so that all connections are properly removed when a
port's type changes, but for now I'll just leave that at "change the
type of a connected port at your own risk".